--- model/stn.py
@@ -1,61 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Jul  8 13:18:36 2019
-
-@author: xingyu
-"""
-
-import torch.nn as nn
-import torch
-import torch.nn.functional as F
-
-
-class STNet(nn.Module):
-
-    def __init__(self):
-        super(STNet, self).__init__()
-
-        # Spatial transformer localization-network
-        self.localization = nn.Sequential(
-            nn.Conv2d(3, 32, kernel_size=3),
-            nn.MaxPool2d(2, stride=2),
-            nn.Mish(True),
-            nn.Conv2d(32, 32, kernel_size=5),
-            nn.MaxPool2d(3, stride=3),
-            nn.Mish(True)
-        )
-        # Regressor for the 3x2 affine matrix
-        self.fc_loc = nn.Sequential(
-            nn.Linear(32 * 15 * 6, 32),
-            nn.Mish(True),
-            nn.Linear(32, 3 * 2)
-        )
-        # Initialize the weights/bias with identity transformation 
-        self.fc_loc[2].weight.data.zero_()
-        self.fc_loc[2].bias.data.copy_(torch.tensor([1, 0, 0, 0, 1, 0], dtype=torch.float))
-
-    def forward(self, x):
-        xs = self.localization(x)
-        xs = xs.view(-1, 32 * 15 * 6)
-        theta = self.fc_loc(xs)
-        theta = theta.view(-1, 2, 3)
-
-        grid = F.affine_grid(theta, x.size(), align_corners=True)
-        x = F.grid_sample(x, grid, align_corners=True)
-
-        return x
-
-
-if __name__ == "__main__":
-    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    model = STNet().to(device)
-
-    input = torch.Tensor(2, 3, 50, 100).to(device)
-    output = model(input)
-    print('output shape is', output.shape)
-
-    from torchsummary import summary
-
-    summary(model, (3, 50, 100), device='cuda')
-
