+++ yaml_test.py
@@ -1,0 +1,6 @@
+import yaml
+from argparse import Namespace
+
+with open('args.yaml') as f:
+    args = Namespace(**yaml.load(f, Loader=yaml.FullLoader))
+
